pyproject.toml
[tool.poetry]
name = "tsp-solver"
version = "0.1.0"
description = ""
authors = ["Larkin Mikhail <jetpilot1917@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.29"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
alembic = "^1.13.1"
black = "^24.3.0"
ruff = "^0.3.5"
alembic-postgresql-enum = "^1.1.2"
aio-pika = "^9.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
